+++ tests/nouns/adjectives.test.js
@@ -1,0 +1,17 @@
+const test = require('tape')
+const nlp = require('../_lib')
+
+test('.adjectives():', function(t) {
+  let doc = nlp('the really cute cat')
+  let m = doc.nouns().adjectives()
+  t.equal(m.text(), 'cute', 'cute .')
+
+  doc = nlp('the really cute orange cat')
+  m = doc.nouns().adjectives()
+  t.equal(m.text(), 'cute orange', 'two adjectives')
+
+  doc = nlp('the cat who was really mean')
+  m = doc.nouns().adjectives()
+  t.equal(m.text(), 'mean', 'who was really .')
+  t.end()
+})